other(refactoring): Update the document module structure+ singleton InMemomyDocumentStore
