+++ connector-runtime/connector-runtime-core/src/main/java/io/camunda/connector/runtime/core/inbound/InboundConnectorContextImpl.java
@@ -32,10 +32,10 @@ import io.camunda.connector.runtime.core.inbound.correlation.InboundCorrelationH
 import io.camunda.connector.runtime.core.inbound.details.InboundConnectorDetails;
 import io.camunda.connector.runtime.core.inbound.details.InboundConnectorDetails.ValidInboundConnectorDetails;
 import io.camunda.document.Document;
-import io.camunda.document.DocumentFactory;
-import io.camunda.document.DocumentFactoryImpl;
-import io.camunda.document.InMemoryDocumentStore;
+import io.camunda.document.factory.DocumentFactory;
+import io.camunda.document.factory.DocumentFactoryImpl;
 import io.camunda.document.store.DocumentCreationRequest;
+import io.camunda.document.store.InMemoryDocumentStore;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -56,12 +56,10 @@ public class InboundConnectorContextImpl extends AbstractConnectorContext
   private final ObjectMapper objectMapper;
 
   private final Consumer<Throwable> cancellationCallback;
-
-  private Health health = Health.unknown();
-
   private final EvictingQueue<Activity> logs;
-
   private final DocumentFactory documentFactory;
+  private Health health = Health.unknown();
+  private Map<String, Object> propertiesWithSecrets;
 
   public InboundConnectorContextImpl(
       SecretProvider secretProvider,
@@ -95,7 +93,7 @@ public class InboundConnectorContextImpl extends AbstractConnectorContext
     this(
         secretProvider,
         validationProvider,
-        new DocumentFactoryImpl(new InMemoryDocumentStore()),
+        new DocumentFactoryImpl(InMemoryDocumentStore.INSTANCE),
         connectorDetails,
         correlationHandler,
         cancellationCallback,
@@ -181,8 +179,6 @@ public class InboundConnectorContextImpl extends AbstractConnectorContext
   public Document createDocument(DocumentCreationRequest request) {
     return documentFactory.create(request);
   }
-
-  private Map<String, Object> propertiesWithSecrets;
 
   private Map<String, Object> getPropertiesWithSecrets(Map<String, Object> properties) {
     if (propertiesWithSecrets == null) {
