+++ connector-sdk/document/src/main/java/io/camunda/document/operation/AggregatingOperationExecutor.java
@@ -14,18 +14,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.document;
+package io.camunda.document.operation;
 
-import io.camunda.document.store.DocumentCreationRequest;
+import io.camunda.document.Document;
 
-public interface DocumentFactory {
+public class AggregatingOperationExecutor implements DocumentOperationExecutor {
 
-  /** Given a document reference, create the Document object */
-  Document resolve(DocumentReference reference);
+  public AggregatingOperationExecutor() {}
 
-  /**
-   * Upload a document to the underlying document store and parse the document reference into a
-   * Document object
-   */
-  Document create(DocumentCreationRequest request);
+  @Override
+  public boolean matches(DocumentOperation operationReference) {
+    return true;
+  }
+
+  @Override
+  public Object execute(DocumentOperation operationReference, Document document) {
+    return null;
+  }
 }
