+++ connector-sdk/document/src/main/java/io/camunda/document/store/CamundaDocumentStore.java
@@ -16,7 +16,7 @@
  */
 package io.camunda.document.store;
 
-import io.camunda.document.DocumentReference.CamundaDocumentReference;
+import io.camunda.document.reference.DocumentReference.CamundaDocumentReference;
 import java.io.InputStream;
 
 public interface CamundaDocumentStore {
