+++ connector-sdk/jackson-datatype-document/src/main/java/io/camunda/connector/document/annotation/jackson/deserializer/DocumentDeserializerBase.java
@@ -23,7 +23,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 import io.camunda.connector.document.annotation.jackson.DocumentOperationResult;
 import io.camunda.connector.document.annotation.jackson.DocumentReferenceModel;
 import io.camunda.document.Document;
-import io.camunda.document.DocumentFactory;
+import io.camunda.document.factory.DocumentFactory;
 import io.camunda.document.operation.DocumentOperation;
 import io.camunda.document.operation.DocumentOperationExecutor;
 import java.io.IOException;
