+++ connector-sdk/jackson-datatype-document/src/main/java/io/camunda/connector/document/annotation/jackson/deserializer/ObjectDocumentDeserializer.java
@@ -20,7 +20,7 @@ import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.deser.std.UntypedObjectDeserializer;
 import io.camunda.connector.document.annotation.jackson.DocumentReferenceModel;
-import io.camunda.document.DocumentFactory;
+import io.camunda.document.factory.DocumentFactory;
 import io.camunda.document.operation.DocumentOperationExecutor;
 import java.io.IOException;
 import java.util.ArrayList;
@@ -30,14 +30,13 @@ public class ObjectDocumentDeserializer extends DocumentDeserializerBase<Object>
 
   private final UntypedObjectDeserializer fallbackDeserializer =
       new UntypedObjectDeserializer(null, null);
+  private final boolean lazy;
 
   public ObjectDocumentDeserializer(
       DocumentOperationExecutor operationExecutor, DocumentFactory documentFactory, boolean lazy) {
     super(operationExecutor, documentFactory);
     this.lazy = lazy;
   }
-
-  private final boolean lazy;
 
   @Override
   public Object deserializeDocumentReference(
