+++ connector-sdk/jackson-datatype-document/src/test/java/io/camunda/connector/document/jackson/DocumentSerializationTest.java
@@ -25,7 +25,7 @@ import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
 import io.camunda.connector.document.annotation.jackson.DocumentReferenceModel.CamundaDocumentReferenceModel;
 import io.camunda.connector.document.annotation.jackson.JacksonModuleDocument;
 import io.camunda.document.Document;
-import io.camunda.document.DocumentFactory;
+import io.camunda.document.factory.DocumentFactory;
 import io.camunda.document.operation.DocumentOperationExecutor;
 import java.util.Map;
 import java.util.Optional;
@@ -43,8 +43,6 @@ public class DocumentSerializationTest {
       new ObjectMapper()
           .registerModule(new JacksonModuleDocument(factory, operationExecutor))
           .registerModule(new Jdk8Module());
-
-  record SourceTypeDocument(Document document) {}
 
   @Test
   void sourceTypeDocument() throws JsonProcessingException, JSONException {
@@ -66,4 +64,6 @@ public class DocumentSerializationTest {
         """;
     JSONAssert.assertEquals(expectedResult, result, true);
   }
+
+  record SourceTypeDocument(Document document) {}
 }
